implement update product

+++ src/modules/product/commands/delete/delete-product.service.ts
@@ -1,5 +1,5 @@
 import { ProductRepositoryPort } from '../../database/product.repository.port';
-import { Inject } from '@nestjs/common';
+import { Inject, NotFoundException } from '@nestjs/common';
 import { CommandHandler } from '@nestjs/cqrs';
 import { DeleteProductCommand } from './delete-product.command';
 
@@ -10,8 +10,13 @@ export class DeleteProductService {
     private readonly productRepo: ProductRepositoryPort,
   ) {}
 
-  async execute(command: DeleteProductCommand): Promise<void> {
-    this.productRepo.findOneByIdOrThrow(command.id);
-    this.productRepo.delete(command.id);
+  async execute(
+    command: DeleteProductCommand,
+  ): Promise<Record<string, boolean | number>> {
+    if (await this.productRepo.exists({ _id: command.id })) {
+      return await this.productRepo.delete(command.id);
+    } else {
+      throw new NotFoundException(`${command.id} does not exist`);
+    }
   }
 }
